chore: comment, refactor, and reduce public api
